Cart changes and interaction with the server
Add downloading of complete selected products from localStorage. If the cart is empty, block sending the cart to the server. After sending the list of products to the server, empty the cart. Open cart on page load works fine. Remove focus from main page when opening modal window

Add a message about the successful sending of goods from the cart to the server and the time of receipt.

Improve code readability.